Update beta1 weebhook, kustomization.yaml and go version
