+++ Makefile
@@ -143,13 +143,6 @@ generate-go:
 		--output-file-base=zz_generated.conversion $(GEN_OUTPUT_BASE) \
 		--go-header-file=./hack/boilerplate.go.txt
 
-	$(CONVERSION_GEN) \
-		--input-dirs=./api/v1alpha4 \
-		--extra-peer-dirs=github.com/spectrocloud/cluster-api-provider-maas/api/v1alpha4 \
-		--build-tag=ignore_autogenerated_core_v1alpha4 \
-		--output-file-base=zz_generated.conversion $(GEN_OUTPUT_BASE) \
-		--go-header-file=./hack/boilerplate.go.txt
-
 generate-manifests:  ## Generate manifests
 	$(CONTROLLER_GEN) $(CRD_OPTIONS) rbac:roleName=manager-role webhook paths="./..." output:crd:artifacts:config=config/crd/bases
