+++ api/v1alpha3/webhook_suite_test.go
@@ -22,7 +22,7 @@ import (
 	"path"
 	"testing"
 
-	infrav1 "github.com/spectrocloud/cluster-api-provider-maas/api/v1alpha4"
+	infrav1 "github.com/spectrocloud/cluster-api-provider-maas/api/v1beta1"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
