+++ api/v1alpha4/maascluster_types.go
@@ -82,7 +82,7 @@ func (in APIEndpoint) IsZero() bool {
 	return in.Host == "" && in.Port == 0
 }
 
-// +kubebuilder:resource:path=maasmachines,scope=Namespaced,categories=cluster-api
+// +kubebuilder:resource:path=maasclusters,scope=Namespaced,categories=cluster-api
 // +kubebuilder:object:root=true
 // +kubebuilder:subresource:status
