+++ api/v1beta1/maasmachinetemplate_types.go
@@ -31,6 +31,7 @@ type MaasMachineTemplateResource struct {
 	Spec MaasMachineSpec `json:"spec"`
 }
 
+// +kubebuilder:resource:path=maasmachinetemplates,scope=Namespaced,categories=cluster-api
 //+kubebuilder:object:root=true
 //+kubebuilder:subresource:status
 //+kubebuilder:storageversion
