+++ config/crd/bases/infrastructure.cluster.x-k8s.io_maasmachines.yaml
@@ -314,7 +314,7 @@ spec:
             type: object
         type: object
     served: true
-    storage: true
+    storage: false
     subresources:
       status: {}
   - name: v1beta1
@@ -467,7 +467,7 @@ spec:
             type: object
         type: object
     served: true
-    storage: false
+    storage: true
     subresources:
       status: {}
 status:
