+++ config/crd/kustomization.yaml
@@ -1,6 +1,7 @@
 commonLabels:
   cluster.x-k8s.io/v1alpha3: v1alpha3
   cluster.x-k8s.io/v1alpha4: v1alpha4
+  cluster.x-k8s.io/v1beta1: v1beta1
 
 # This kustomization.yaml is not intended to be run by itself,
 # since it depends on service name and namespace that are out of this kustomize package.
