+++ metadata.yaml
@@ -11,3 +11,6 @@ releaseSeries:
   - major: 0
     minor: 2
     contract: v1alpha4
+  - major: 1
+    minor: 1
+    contract: v1beta1
